requirements.txt
fontbakery
arabic-reshaper
python-bidi